organize api.builder module
